requirements.txt
tweepy>=4.14.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
schedule>=1.2.0
loguru>=0.7.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
